refactor: move StateList component to separate file
