+++ src/pages/data/index.js
@@ -1,39 +1,12 @@
 import React from 'react'
 import { graphql, Link } from 'gatsby'
-import { Flex, Box } from '../../components/common/flexbox'
-import State from '../../components/pages/data/state-data'
+import { Box } from '../../components/common/flexbox'
+import StateList from '../../components/pages/data/state-list'
 import Layout from '../../components/layout'
 import { SyncInfobox } from '../../components/common/infobox'
 import DetailText from '../../components/common/detail-text'
 import SummaryTable from '../../components/common/summary-table'
 import '../../scss/pages/data.scss'
-
-const StateList = ({ states, stateData }) => {
-  const stateList = []
-  states.forEach(({ node }) => {
-    const state = node
-    stateData.forEach(data => {
-      if (data.node.state === state.state) {
-        state.stateData = data.node
-      }
-    })
-    stateList.push(state)
-  })
-  return (
-    <Flex flexWrap="wrap" m="0 -10px">
-      {stateList.map(state => (
-        <Box
-          width={1}
-          mb={['1rem', '1.5rem']}
-          p="0 10px"
-          className="data-state"
-        >
-          <State state={state} stateData={state.stateData} />
-        </Box>
-      ))}
-    </Flex>
-  )
-}
 
 const StatesNav = ({ stateList }) => (
   <Box width={[1, 1, 1, 1 / 2]} className="state-nav" m="0 auto 1rem">
